:sparkles: move market to agent
